third_party/blink/web_tests/platform/generic/http/tests/security/cross-frame-access-delete-expected.txt

PASS: deletionThrowsException('targetWindow.existingProperty') should be 'true' and is.
PASS: deletionThrowsException('targetWindow[1]') should be 'true' and is.
PASS: deletionThrowsException('targetWindow.location.existingProperty') should be 'true' and is.
PASS: deletionThrowsException('targetWindow.location[1]') should be 'true' and is.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------

----- tests for deleting the properties of window, window.history, and window.location cross-domain -----

PASS: window.existingProperty should be 'test value' and is.
PASS: window[1] should be 'undefined' and is.
PASS: window.history.existingProperty should be 'test value' and is.
PASS: window.history[1] should be 'test value' and is.
PASS: window.location.existingProperty should be 'test value' and is.
PASS: window.location[1] should be 'test value' and is.
